requirements.txt
# Core Agent & Graph
langchain
langgraph
ollama
pydantic

# Utilities
loguru
python-dotenv
requests
beautifulsoup4
